+++ pkgs/ffigen/test/collision_tests/expected_bindings/_expected_decl_decl_collision_bindings.dart
@@ -65,8 +65,6 @@ const int Test_Macro1 = 0;
 
 typedef testAlias = ffi.Void;
 typedef DarttestAlias = void;
-typedef testAlias1 = ffi.Void;
-typedef DarttestAlias1 = void;
 
 final class testCrossDecl extends ffi.Opaque {}